float pipeline: forbid FMA contraction so bit-exactness holds on every GOARCH
Go's spec permits the compiler to contract `x*y + z` into a single-rounding
FMA, and gc does so on arm64, ppc64, s390x, riscv64 and loong64 — but never on
amd64. The oracle vectors come from an x86-64 SSE build that rounds each
product separately, so the float32 kernels silently diverged from the
bit-for-bit contract on those architectures while the amd64 gates stayed green.
Compiling mct.DecodeReal for GOARCH=arm64 emitted FMADDS.

Wrap every float product that feeds an add or a subtract in an explicit
float32(...)/float64(...) conversion. The spec defines such a conversion as a
rounding step, which makes it a guaranteed fusion barrier; it forbids fusion
only, so operand order and associativity are untouched — including the
deliberate gcc -freassoc groupings in mct.EncodeReal/DecodeReal that reproduce
the shipped libopenjp2.so. On amd64 the emitted code is unchanged and all
oracle gates remain byte-identical.

Sites covered: the 9/7 lifting kernels (dwt97.go), the irreversible MCT and the
custom-MCT matrix multiply and inversion (mct.go, invert.go), the float64 norm
accumulation feeding rate allocation (CalculateNorms, tcd maxSE), and the
sYCC/CMYK/eYCC/CIELab colour conversions, which the asm showed were float
rather than integer.

Three FMA sites remain in repo code and are justified in place: the 11-bit
mantissa scale in init_tile.go and the halving in the rate-allocation bisection
are both exact, so fusion is a semantic no-op there, and the FMSUBD on
riscv64/ppc64le/loong64 in j2k_encode_write.go belongs to the compiler's own
float64->uint64 conversion lowering.

Verified: zero FMADD/FMSUB/FNMADD/FNMSUB in dwt, mct, j2k and the root package
for arm64, ppc64le, s390x, riscv64 and loong64; amd64 instruction streams diff
clean apart from one commutative ADDSS operand swap; full oracletest suite
green with identical CIELab diff counts; dwt and whole-image decode benchmarks
unchanged within noise.

+++ internal/dwt/dwt97.go
@@ -4,6 +4,16 @@ package dwt
 // 1-D lifting kernels, the forward/inverse row and column passes, and the
 // whole-tile 2-D drivers. All arithmetic is float32 in the exact order of the
 // C source; results are bit-exact with the C library.
+//
+// FMA barriers: Go's spec permits the compiler to fuse `x*y + z` (and the
+// `x*y - z` / `z - x*y` shapes) into a single-rounding FMA, and gc does so on
+// arm64/ppc64/s390x/riscv64. The C reference rounds the product separately, so
+// every float32 product that feeds an add or a subtract below is wrapped in an
+// explicit float32(...) conversion. Per the Go spec an explicit floating-point
+// conversion rounds to the target type's precision and therefore forbids the
+// fusion; on amd64 (where gc never fuses) it is a no-op, so the byte-identical
+// oracle parity on that arch is unchanged. The conversions only forbid fusion —
+// operand order and associativity are exactly as before.
 
 // two_invK: historic value for 2 / opj_invK used by the inverse transform
 // (see the BUG_WEIRD_TWO_INVK note in dwt.c / tcd.c).
@@ -16,15 +26,15 @@ const dwtTwoInvK float32 = 1.625732422
 func encodeStep2(w []float32, flOff, fwOff int, end, m uint32, c float32) {
 	imax := uintMin(end, m)
 	if imax > 0 {
-		w[fwOff-1] += (w[flOff] + w[fwOff]) * c
+		w[fwOff-1] += float32((w[flOff] + w[fwOff]) * c)
 		fwOff += 2
 		for i := uint32(1); i < imax; i++ {
-			w[fwOff-1] += (w[fwOff-2] + w[fwOff]) * c
+			w[fwOff-1] += float32((w[fwOff-2] + w[fwOff]) * c)
 			fwOff += 2
 		}
 	}
 	if m < end {
-		w[fwOff-1] += (2 * w[fwOff-2]) * c
+		w[fwOff-1] += float32((2 * w[fwOff-2]) * c)
 	}
 }
 
@@ -151,19 +161,19 @@ func encodeV8Step2(tmp []float32, flOff, fwOff int, end, m uint32, cst float32) 
 	imax := uintMin(end, m)
 	if imax > 0 {
 		for c := 0; c < nbEltsV8; c++ {
-			tmp[fwOff-nbEltsV8+c] += (tmp[flOff+c] + tmp[fwOff+c]) * cst
+			tmp[fwOff-nbEltsV8+c] += float32((tmp[flOff+c] + tmp[fwOff+c]) * cst)
 		}
 		fwOff += 2 * nbEltsV8
 		for i := uint32(1); i < imax; i++ {
 			for c := 0; c < nbEltsV8; c++ {
-				tmp[fwOff-nbEltsV8+c] += (tmp[fwOff-2*nbEltsV8+c] + tmp[fwOff+c]) * cst
+				tmp[fwOff-nbEltsV8+c] += float32((tmp[fwOff-2*nbEltsV8+c] + tmp[fwOff+c]) * cst)
 			}
 			fwOff += 2 * nbEltsV8
 		}
 	}
 	if m < end {
 		for c := 0; c < nbEltsV8; c++ {
-			tmp[fwOff-nbEltsV8+c] += (2 * tmp[fwOff-2*nbEltsV8+c]) * cst
+			tmp[fwOff-nbEltsV8+c] += float32((2 * tmp[fwOff-2*nbEltsV8+c]) * cst)
 		}
 	}
 }
@@ -286,7 +296,7 @@ func v8dwtDecodeStep2(w []float32, lUnitOff, wUnitOff int, start, end, m uint32,
 	}
 	for i := start; i < imax; i++ {
 		for e := 0; e < nbEltsV8; e++ {
-			w[wpos-nbEltsV8+e] = w[wpos-nbEltsV8+e] + (w[lpos+e]+w[wpos+e])*c
+			w[wpos-nbEltsV8+e] = w[wpos-nbEltsV8+e] + float32((w[lpos+e]+w[wpos+e])*c)
 		}
 		lpos = wpos
 		wpos += 2 * nbEltsV8
@@ -294,7 +304,7 @@ func v8dwtDecodeStep2(w []float32, lUnitOff, wUnitOff int, start, end, m uint32,
 	if m < end {
 		c2 := c + c
 		for e := 0; e < nbEltsV8; e++ {
-			w[wpos-nbEltsV8+e] = w[wpos-nbEltsV8+e] + w[lpos+e]*c2
+			w[wpos-nbEltsV8+e] = w[wpos-nbEltsV8+e] + float32(w[lpos+e]*c2)
 		}
 	}
 }
